Backend Refactor Completed
