+++ Middleware/verifyJWT.js
@@ -1,0 +1,21 @@
+import jwt from "jsonwebtoken";
+
+export const verifyJWT = (req, res, next) => {
+  const authHeader = req.headers["authorization"];
+
+  const token = authHeader && authHeader.split(" ")[1];
+
+  if (!token) {
+    return res.status(401).json({ error: "No Access Token" });
+  }
+
+  jwt.verify(token, process.env.SECRET_ACCESS_KEY, (err, user) => {
+    if (err) {
+      console.log(err);
+      return res.status(403).json({ err: "Access Token is Invalid" });
+    }
+
+    req.user = user.id;
+    next();
+  });
+};
